Fix error when version not found
The closeChangelog function was throwing an error when the newly
"closed" version was not found in the changelog.  There were two
issues, the version property of a changelog version has any leading
'v' removed, and the regular expression used to extract the semver is
not very good, so sometimes it is completely wrong.  This led to the
version not being found, which led to the error because the code was
not ensuring a version was found.  I fixed the first issue by matching
against the link text in the title and the second issue by checking
the return of the find.  Better leverage types.  Add tests.

[changelog:fixed]